+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add `replace_Inf()` to replace infinite values in a `matrix`-like object.
 * Add `replace_zero()` to replace zero in a `matrix`-like object.
 * Add `remove_Inf()` to remove infinite values in a `matrix`-like object.
+* Add resampling methods: `bootstrap()` and `jackknife()`.
 * Add `assert_*()` and `validate()` to validate objects.
 * Add `get_dates()` and `set_dates()<-` to extract/replace `dates_from` and `dates_to` slots.